+++ app/src/main/java/woowacourse/shopping/data/cart/Cart.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.model.db.cart
+package woowacourse.shopping.data.cart
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
