+++ app/src/main/java/woowacourse/shopping/data/cart/CartDao.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.model.db.cart
+package woowacourse.shopping.data.cart
 
 import androidx.room.Dao
 import androidx.room.Insert
