+++ app/src/main/java/woowacourse/shopping/data/cart/CartDatabase.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.model.db.recentproduct
+package woowacourse.shopping.data.cart
 
 import android.content.Context
 import androidx.room.Database
@@ -6,17 +6,17 @@ import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
 
-@Database(entities = [RecentProduct::class], version = 1)
-@TypeConverters(RecentProductTypeConverters::class)
-abstract class RecentProductDatabase : RoomDatabase() {
-    abstract fun recentProductDao(): RecentProductDao
+@Database(entities = [Cart::class], version = 1)
+@TypeConverters(CartTypeConverters::class)
+abstract class CartDatabase : RoomDatabase() {
+    abstract fun cartDao(): CartDao
 
     companion object {
-        private const val DATABASE_NAME = "recent_product_db"
+        private const val DATABASE_NAME = "cart_db"
         private const val ERROR_DATABASE = "데이터베이스가 초기화 되지 않았습니다."
-        private var database: RecentProductDatabase? = null
+        private var database: CartDatabase? = null
 
-        fun database(): RecentProductDatabase {
+        fun database(): CartDatabase {
             return database ?: throw IllegalStateException(ERROR_DATABASE)
         }
 
@@ -25,7 +25,7 @@ abstract class RecentProductDatabase : RoomDatabase() {
                 database =
                     Room.databaseBuilder(
                         context.applicationContext,
-                        RecentProductDatabase::class.java,
+                        CartDatabase::class.java,
                         DATABASE_NAME,
                     ).build()
             }
