+++ app/src/main/java/woowacourse/shopping/data/product/ProductDao.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.model.data
+package woowacourse.shopping.data.product
 
 import woowacourse.shopping.model.Product
