+++ app/src/main/java/woowacourse/shopping/data/recentproduct/RecentProduct.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.model.db.recentproduct
+package woowacourse.shopping.data.recentproduct
 
 import androidx.room.Entity
 import androidx.room.Index
