+++ app/src/main/java/woowacourse/shopping/data/recentproduct/RecentProductRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.model.db.recentproduct
+package woowacourse.shopping.data.recentproduct
 
 import java.time.LocalDateTime
