+++ app/src/main/java/woowacourse/shopping/data/recentproduct/RecentProductTypeConverters.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.model.db.recentproduct
+package woowacourse.shopping.data.recentproduct
 
 import androidx.room.TypeConverter
 import java.time.LocalDateTime
