+++ app/src/main/java/woowacourse/shopping/ui/detail/ProductDetailActivity.kt
@@ -9,12 +9,12 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import woowacourse.shopping.R
+import woowacourse.shopping.data.cart.CartDatabase
+import woowacourse.shopping.data.cart.CartRepositoryImpl
+import woowacourse.shopping.data.product.ProductsImpl
+import woowacourse.shopping.data.recentproduct.RecentProductDatabase
+import woowacourse.shopping.data.recentproduct.RecentProductRepositoryImpl
 import woowacourse.shopping.databinding.ActivityProductDetailBinding
-import woowacourse.shopping.model.data.ProductsImpl
-import woowacourse.shopping.model.db.cart.CartDatabase
-import woowacourse.shopping.model.db.cart.CartRepositoryImpl
-import woowacourse.shopping.model.db.recentproduct.RecentProductDatabase
-import woowacourse.shopping.model.db.recentproduct.RecentProductRepositoryImpl
 import woowacourse.shopping.ui.CountButtonClickListener
 import woowacourse.shopping.ui.detail.viewmodel.ProductDetailViewModel
 import woowacourse.shopping.ui.detail.viewmodel.ProductDetailViewModelFactory
