+++ app/src/main/java/woowacourse/shopping/ui/detail/viewmodel/ProductDetailViewModel.kt
@@ -4,11 +4,11 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.map
+import woowacourse.shopping.data.cart.CartRepository
+import woowacourse.shopping.data.product.ProductDao
+import woowacourse.shopping.data.recentproduct.RecentProductRepository
 import woowacourse.shopping.model.Product
 import woowacourse.shopping.model.ProductWithQuantity
-import woowacourse.shopping.model.data.ProductDao
-import woowacourse.shopping.model.db.cart.CartRepository
-import woowacourse.shopping.model.db.recentproduct.RecentProductRepository
 import woowacourse.shopping.ui.utils.Event
 import kotlin.concurrent.thread
