+++ app/src/main/java/woowacourse/shopping/ui/detail/viewmodel/ProductDetailViewModelFactory.kt
@@ -2,9 +2,9 @@ package woowacourse.shopping.ui.detail.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import woowacourse.shopping.model.data.ProductDao
-import woowacourse.shopping.model.db.cart.CartRepository
-import woowacourse.shopping.model.db.recentproduct.RecentProductRepository
+import woowacourse.shopping.data.cart.CartRepository
+import woowacourse.shopping.data.product.ProductDao
+import woowacourse.shopping.data.recentproduct.RecentProductRepository
 
 class ProductDetailViewModelFactory(
     private val productDao: ProductDao,
