+++ app/src/main/java/woowacourse/shopping/ui/products/viewmodel/ProductContentsViewModelFactory.kt
@@ -2,9 +2,9 @@ package woowacourse.shopping.ui.products.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import woowacourse.shopping.model.data.ProductDao
-import woowacourse.shopping.model.db.cart.CartRepository
-import woowacourse.shopping.model.db.recentproduct.RecentProductRepository
+import woowacourse.shopping.data.cart.CartRepository
+import woowacourse.shopping.data.product.ProductDao
+import woowacourse.shopping.data.recentproduct.RecentProductRepository
 
 class ProductContentsViewModelFactory(
     private val productDao: ProductDao,
